+++ src/main/java/com/bloomtech/asylumrgbea/mappers/AsylumCaseMapper.java
@@ -1,8 +1,8 @@
 package com.bloomtech.asylumrgbea.mappers;
 
 import com.bloomtech.asylumrgbea.entities.AsylumCase;
-import com.bloomtech.asylumrgbea.models.AsylumCaseRequestDto;
-import com.bloomtech.asylumrgbea.models.AsylumCaseResponseDto;
+import com.bloomtech.asylumrgbea.models.CaseResponseDto;
+import com.bloomtech.asylumrgbea.models.CasesRequestDto;
 import com.bloomtech.asylumrgbea.models.PageResponseDto;
 import org.mapstruct.Mapper;
 import org.springframework.data.domain.Page;
@@ -13,13 +13,13 @@ import org.springframework.data.domain.Page;
 @Mapper(uses = {AsylumCaseMapper.class}, componentModel = "spring")
 public interface AsylumCaseMapper {
 
-    AsylumCase requestToEntity(AsylumCaseRequestDto asylumCaseRequestDto);
+    AsylumCase requestToEntity(CasesRequestDto casesRequestDto);
 
-    AsylumCaseResponseDto entityToResponseDto(AsylumCase asylumCase);
+    CaseResponseDto entityToResponseDto(AsylumCase asylumCase);
 
-    Iterable<AsylumCaseResponseDto> entitiesToResponseDtos(Iterable<AsylumCase> asylumCases);
+    Iterable<CasesRequestDto> entitiesToResponseDtos(Iterable<AsylumCase> asylumCases);
 
-    Iterable<AsylumCaseResponseDto> pagesToResponseDtos(Page<AsylumCase> asylumCases);
+    Iterable<CaseResponseDto> pagesToResponseDtos(Page<AsylumCase> asylumCases);
 
-    PageResponseDto pageDataAndPageToResponseDto(int totalPages, Iterable<AsylumCaseResponseDto> page);
+    PageResponseDto pageDataAndPageToResponseDto(int totalPages, Iterable<CaseResponseDto> page);
 }
