--- src/main/java/com/bloomtech/asylumrgbea/models/AsylumCaseRequestDto.java
@@ -1,17 +1,0 @@
-package com.bloomtech.asylumrgbea.models;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-@AllArgsConstructor
-public class AsylumCaseRequestDto {
-
-    private Integer numberOfItemsInPage;
-
-    private int pageNumber;
-}
