+++ src/main/java/com/bloomtech/asylumrgbea/models/PageResponseDto.java
@@ -14,5 +14,5 @@ public class PageResponseDto {
 
     private int totalPages;
 
-    private Iterable<AsylumCaseResponseDto> page;
+    private Iterable<CaseResponseDto> page;
 }
