+++ src/main/java/com/bloomtech/asylumrgbea/services/AsylumCaseService.java
@@ -1,13 +1,62 @@
 package com.bloomtech.asylumrgbea.services;
 
-import com.bloomtech.asylumrgbea.models.AsylumCaseRequestDto;
-import com.bloomtech.asylumrgbea.models.AsylumCaseResponseDto;
-import com.bloomtech.asylumrgbea.models.PageResponseDto;
+import com.bloomtech.asylumrgbea.controllers.exceptions.AsylumCaseNotFoundException;
+import com.bloomtech.asylumrgbea.controllers.exceptions.BadRequestException;
+import com.bloomtech.asylumrgbea.entities.AsylumCase;
+import com.bloomtech.asylumrgbea.mappers.AsylumCaseMapper;
+import com.bloomtech.asylumrgbea.models.CasesRequestDto;
+import com.bloomtech.asylumrgbea.repositories.AsylumCaseRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.cache.annotation.Cacheable;
+import org.springframework.stereotype.Service;
 
-public interface AsylumCaseService {
+@Service
+@RequiredArgsConstructor
+public class AsylumCaseService {
 
-	PageResponseDto getPageOfAsylumCases(AsylumCaseRequestDto asylumCaseRequestDto);
+	private final AsylumCaseRepository asylumCaseRepository;
 
-	Iterable<AsylumCaseResponseDto> getAllAsylumCases();
+	private final AsylumCaseMapper asylumCaseMapper;
 
+	/**
+	 * Gets all entities from repository and returns them out of the program.
+	 * @return This is an Iterable of AsylumCase objects.
+	 */
+	@Cacheable("asylum_case_cache")
+	public Iterable<CasesRequestDto> getAllAsylumCases() {
+
+		Iterable<AsylumCase> caseIterable = asylumCaseRepository.findAll();
+		validateIterableIsNotEmpty(caseIterable);
+
+		return asylumCaseMapper.entitiesToResponseDtos(caseIterable);
+	}
+
+	/**
+	 * Gets a page of entities from repository and returns them out of the program.
+	 * @param casesRequestDto Contains the page number as a field.
+	 * @return This is a limited Iterable of AsylumCase objects.
+	 */
+
+
+	/**
+	 * Checks if an Iterable has at least 1 element and throws an exception if not so.
+	 * @param caseIterable An empty or non-empty Iterable of AsylumCase objects.
+	 * @throws AsylumCaseNotFoundException if the Iterable is empty this exception is thrown.
+	 */
+	private void validateIterableIsNotEmpty(Iterable<AsylumCase> caseIterable) throws AsylumCaseNotFoundException {
+
+		if (!caseIterable.iterator().hasNext())
+			throw new AsylumCaseNotFoundException("ERROR: No cases were found...");
+	}
+
+	/**
+	 * Checks if a CasesRequestDto object contains a null field and throws an exception if so.
+	 * @param casesRequestDto package and a request Dto.
+	 * @throws BadRequestException The exception thrown if triggered.
+	 */
+	private void validateRequestDto(CasesRequestDto casesRequestDto) throws BadRequestException {
+
+		if (casesRequestDto.getNumberOfItemsInPage() == null)
+			throw new BadRequestException("ERROR: The field number of items in a page cannot be null...");
+	}
 }
